% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{as_tabular}
\alias{as_tabular}
\title{Coerce to tabular}
\usage{
as_tabular(x, class = "data.frame", ...)
}
\arguments{
\item{x}{An R object.}

\item{class}{Desired tabular class (class name or object of that class). If no coercion is necessary, \code{x} is return unchanged.}

\item{...}{Arguments passed to the corresponding coercion function(s).}
}
\description{
Coerces an object to a tabular object of the desired type.
}
\examples{
x <- list(a = 1, b = 2)
as_tabular(x, "data.frame")
\dontrun{
as_tabular(x, "data.table")
as_tabular(x, "tbl_df")
as_tabular(x, "tbl_dt")
}
}
