% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\docType{package}
\name{beta}
\alias{beta}
\title{beta}
\description{
A package containing functions meant to aid the ETL process. This package heavily relies on functions and concepts from the tidyverse, specifically dplyr.
}
\author{
Tyler Riccio \email{tylerriccio8@gmail.com}
}
