% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.notnull.r
\name{check.notnull}
\alias{check.notnull}
\title{Check not-null constraints}
\usage{
check.notnull(x, cols, silent = FALSE)
}
\arguments{
\item{x}{Data frame representing a table of a data base.}

\item{cols}{Names of the colum(s) of \code{x} to which the constraint
applies (vector of type character).}

\item{silent}{Logical. If \code{FALSE}, details on constraint
violations are shown using \code{\link[base]{print}}.}
}
\value{
\code{TRUE} if the check was passed successfully and
  \code{FALSE} otherwise.
}
\description{
Check for violation of a not-null constraint
}
\note{
The function tests for the occurrence of \code{NA} in any of the
  columns of \code{x} specified in \code{cols}. Hence, it is assumed that
  missing values are marked as \code{NA} (instead of \code{NULL} like in
  a true SQL data base).
}
\examples{

data(temperature)
print(temperature)

# Should succeed
check.notnull(temperature, c("day", "city"))

# Example of NA
temperature[2, "day"] <- NA
print(temperature)
check.notnull(temperature, c("day", "city"))
}
\seealso{
\code{\link{check.unique}}, \code{\link{check.primary}},
  \code{\link{check.foreign}},
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
