% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.unique.r
\name{check.unique}
\alias{check.unique}
\title{Check unique constraints}
\usage{
check.unique(x, cols, silent = FALSE)
}
\arguments{
\item{x}{Data frame representing a table of a data base.}

\item{cols}{Names of the colum(s) of \code{x} to which the constraint
applies (vector of type character).}

\item{silent}{Logical. If \code{FALSE}, details on constraint
violations are shown using \code{\link[base]{print}}.}
}
\value{
\code{TRUE} if the check was passed successfully and
  \code{FALSE} otherwise.
}
\description{
Check for violation of a unique constraint
}
\note{
See examples.
}
\examples{

data(temperature)
print(temperature)

# Should succeed
check.unique(temperature, c("day", "city"))

# Example of duplicates
temperature[2, "day"] <- 1
print(temperature)
check.unique(temperature, c("day", "city"))
}
\seealso{
\code{\link{check.notnull}}, \code{\link{check.primary}},
  \code{\link{check.foreign}}
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
