% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\docType{data}
\name{Classes}
\alias{Classes}
\title{Tabular class definitions}
\format{An object of class \code{list} of length 4.}
\usage{
Classes
}
\description{
Tabular class definitions
}
\keyword{datasets}
