% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{cleaner}
\alias{cleaner}
\title{cleaner}
\usage{
cleaner(.data, .name_fix = T, .nullify = T, .rm_na = T, .dedup = T)
}
\arguments{
\item{.data}{Data frame object to clean}
}
\description{
Sets implied NA values with explicit ones, removes rows or columns containing all NA values, dedups, and converts the column names to a more workable syntax.
}
\examples{
a <- c(1, NA, NA, 4)
b <- c(NA, NA, NA, NA)
c <- c(1, NA, 3, 4)
d <- tibble(a, b, `bad syntax ` = c)
}
