% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce2.R
\name{coalesce2}
\alias{coalesce2}
\title{coalesce2}
\usage{
coalesce2(.df, into = NULL, ... = NULL, .suffix = F)
}
\arguments{
\item{.df}{Dataframe object}

\item{into}{Column to coalesce into}

\item{...}{Columns to coalesce}

\item{.suffix}{Coalesce columns with a common suffix}
}
\value{
A dataframe without the coalesced columns
}
\description{
Coalesce columns into a single one while dropping the used columns in the same function.
}
\examples{
a <- c(1, 2, NA, NA)
b <- c(NA, NA, 3, 4)
c <- tibble::tibble(a, b)
coalesce2(.df = c, into = new_col, a, b)
}
