% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.read.r
\name{db.read}
\alias{db.read}
\title{Read a set of text files forming a data base}
\usage{
db.read(dir = ".", ext = "tsv", sep = "\\t")
}
\arguments{
\item{dir}{Directory containing the text files (character string). Will be
passed to argument \code{path} of \code{\link[base]{list.files}}.}

\item{ext}{A character string representing a file extension used to
identify the files to be read. This is typically a string of three
characters like, e.g., 'csv' or 'tsv'. A preceeding dot is implicitly
assumed to be present and must be omitted.}

\item{sep}{The field delimiter (character) for use with 
\code{\link[base]{read.table}}.}

\item{...}{Further optional arguments passed to \code{\link[base]{read.table}}.}
}
\value{
A list, each element of which is a data frame. Element names are
  constructed from the source file names by stripping the specified extension
  and the preceeding dot.
}
\description{
Reads a set of delimited text files supposed to form a relational data base
}
\note{
The text files are read with \code{\link[base]{read.table}} using the
  fixed arguments \code{header=TRUE} and \code{stringsAsFactors=FALSE}. Thus,
  one should not try to overwrite these settings using the \code{...}
  argument.
}
\examples{

# Read example DB shipped with the package
db <- db.read(dir=system.file("examples", package="tabular"), ext="tsv")
print(names(db))

# Integrity checks wrapped into a dedicated function
validate <- function(db) {
  with(db, {
    stopifnot(check.key(samples, c("date","id_location")))
    stopifnot(check.link(samples, "id_location", locations, "id"))
    # further checks would go here ...
 })
}

validate(db)
}
\seealso{
After reading the set of files, one typically wants to check the
  data base for integrity using the functions \code{\link{check.notnull}},
  \code{\link{check.unique}}, \code{\link{check.key}}, and
  \code{\link{check.link}}. It is probably good style to wrap all necessary
  checks into a single dedicated function that can be called repeatedly
  (e.g. after manipulation of data) or which can be re-used for other
  data bases of the same layout. See example below.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
