% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupe.R
\name{dupe}
\alias{dupe}
\title{dupe}
\usage{
dupe(.df, ...)
}
\arguments{
\item{.df}{Data frame like object}

\item{...}{Column or combination of columns to check}
}
\description{
Adds a column that indicates the input value exists somewhere else in the same column
}
\examples{
a <- c(1, 2, 3, 3)
b <- c(1, 2, 3, 4)
c <- tibble::tibble(a, b)
c \%>\% dupe(a)
}
