% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_phone.R
\name{fix_phone}
\alias{fix_phone}
\title{fix_phone}
\usage{
fix_phone(.df, ..., strip = 11, .format = NULL)
}
\arguments{
\item{.df}{Data frame like object}

\item{...}{Column(s) or vector of strings}

\item{strip}{Number of digits to strip down}

\item{.format}{Optional argument specifying the format}
}
\description{
Formats phone number column or columns
}
\examples{
p <- c("111-111-1111", "2-(222)-222-2222")
p2 <- c("111-111-1111", "2-222-222-2222")
p3 <- tibble::tibble(p, p2)
p3 \%>\% fix_phone(p2, .format = "US")
}
