% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_file.R
\name{in_file}
\alias{in_file}
\title{in_file}
\usage{
in_file(x = NULL, clean = T, choose = F, .assign = F)
}
\arguments{
\item{x}{String or vector of file names}

\item{clean}{Cleans names, removes empty rows, removes constant rows and dedups}

\item{choose}{\code{file.choose()} option}

\item{.assign}{Option to manually declare the variable in the form of a list of tibbles}
}
\description{
Uploads, parses, cleans and assigns in one function
}
\examples{
\dontrun{
in_file(c("example_filename", "filename2"))
}
}
