% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{is_tabular}
\alias{is_tabular}
\title{Test if tabular}
\usage{
is_tabular(x, class = NULL)
}
\arguments{
\item{x}{An R object.}

\item{class}{Tabular class to test against (class name or object of that class) or \code{NULL} to test against all supported classes.}
}
\description{
Tests whether an object is a supported tabular object.
}
\examples{
is_tabular(data.frame()) # TRUE
is_tabular(data.frame(), "data.table") # FALSE
is_tabular(data.frame(), data.frame()) # TRUE
\dontrun{
is_tabular(data.table::data.table()) # TRUE
is_tabular(dtplyr::tbl_dt(data.table::data.table())) # TRUE
is_tabular(tibble::tibble()) # TRUE
}
is_tabular(list()) # FALSE
}
