% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{is_tabular_grid}
\alias{is_tabular_grid}
\title{Test for a \code{tabular_grid} object}
\usage{
is_tabular_grid(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Single logical. \code{TRUE} when \code{x} inherits from \code{tabular_grid},
else \code{FALSE}. Never errors.
}
\description{
Test for a \code{tabular_grid} object
}
\keyword{internal}
