% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listJoin.R
\name{listJoin}
\alias{listJoin}
\title{listJoin}
\usage{
listJoin(x, y, .by)
}
\arguments{
\item{x}{Left data frame object}

\item{y}{Right data frame object}

\item{.by}{Joining column}
}
\value{
Joined table abiding by left join principles.
}
\description{
Left join function that takes a string and searches for a match within a list.
}
\examples{
\dontrun{
firstList <- starwars \%>\%
 select(films) \%>\%
 mutate(films = map(films, ~ .[1]) \%>\% unlist())

firstList \%>\%
 listJoin(starwars \%>\%
            mutate(rowN = row_number())\%>\%
            select(films,rowN) , .by = 'films') \%>\% view()
}
}
