% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeList.R
\name{makeList}
\alias{makeList}
\title{makeList}
\usage{
makeList(data, x, sep = " | ")
}
\arguments{
\item{data}{A data frame object}

\item{x}{Variable}

\item{sep}{String delimeter if applicable}
}
\value{
A tibble with the supplied variable in list format
}
\description{
Vectorized function that turns a string into a delimited list.
}
\examples{
\dontrun{
mpg \%>\%
mutate(manufacturer = paste0(manufacturer, ', something')) \%>\%
makeList(manufacturer, sep = ',')
}
}
