% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cols.R
\name{match_cols}
\alias{match_cols}
\title{match_cols}
\usage{
match_cols(df1, df2)
}
\arguments{
\item{df1}{A dataframe object you want to reorder}

\item{df2}{A data frame object containing the ideal order}
}
\value{
A data frame object with the data from input one in the column order of input 2
}
\description{
Coalesce columns into a single one while dropping the used columns in the same function.
}
\examples{
ideal_order <- tibble::tibble(col_b = NA, col_a = NA)
tibble::tibble(col_a = c(1, 2, 3, 4), col_b = c(5, 6, 7, 8)) \%>\%
  match_cols(ideal_order)
}
