% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(.data, ...)
}
\arguments{
\item{.data}{Data frame object}

\item{...}{Variables to normalize using tidyselect syntax}
}
\description{
Handles skew by square route and
}
\examples{
dplyr::starwars \%>\% normalize(where(is.numeric))
}
