% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{parse_tabular_class}
\alias{parse_tabular_class}
\title{Parse tabular class}
\usage{
parse_tabular_class(x)
}
\arguments{
\item{x}{Character vector of class names or any (non-character) R object.}
}
\description{
Parse tabular class
}
