% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_join.R
\name{soft_join}
\alias{soft_join}
\title{soft_join}
\usage{
soft_join(df1, df2, .by, .fun = dplyr::left_join, .keep = F)
}
\arguments{
\item{df1}{A data frame object}

\item{df2}{A data frame to join}

\item{.by}{A string or character vector delimited with \code{=} as the key(s)}

\item{.fun}{Function to call}

\item{.keep}{Option to keep the joined columns}
}
\value{
A data frame object 1 joined with data frame object 2
}
\description{
Case insensitive join
}
\examples{
a <- starwars \%>\%
  select(1:5)
b <- starwars \%>\%
  select(1:5) \%>\%
  mutate(name = str_to_upper(name)) \%>\%
  mutate(new_column = "hello")

t <- a \%>\%
  multi_join(b, .by = "name")
}
