% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{tabular_predicates}
\alias{tabular_predicates}
\alias{is_tabular_spec}
\alias{is_tabular_grid}
\alias{is_col_spec}
\alias{is_header_node}
\alias{is_sort_spec}
\alias{is_pivot_spec}
\alias{is_derive_spec}
\alias{is_style_node}
\alias{is_style_predicate}
\alias{is_style_spec}
\alias{is_pagination_spec}
\alias{is_preset_spec}
\title{Test for tabular S7 objects}
\usage{
is_tabular_spec(x)

is_tabular_grid(x)

is_col_spec(x)

is_header_node(x)

is_sort_spec(x)

is_pivot_spec(x)

is_derive_spec(x)

is_style_node(x)

is_style_predicate(x)

is_style_spec(x)

is_pagination_spec(x)

is_preset_spec(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Single logical. Never errors.
}
\description{
Predicates returning a single logical indicating whether \code{x}
inherits from the corresponding tabular S7 class.
}
\keyword{internal}
