% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{temperature}
\alias{temperature}
\title{Temperature}
\format{
This is a sample data frame to demonstate table constraints.
}
\description{
A set of temperature data observed at different times and locations.
}
