% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmutate.R
\name{transmutate}
\alias{transmutate}
\title{transmutate}
\usage{
transmutate(.df, ...)
}
\arguments{
\item{.df}{A data frame object you want to reorder}

\item{...}{Parameters to pass to \code{dplyr::mutate}}
}
\value{
A data frame object with the new column in the first position
}
\description{
Mutate columns while dropping the used ones
}
\examples{
dplyr::starwars \%>\%
  transmutate(new_col = height + mass)
}
