% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid.R
\name{uid}
\alias{uid}
\title{uid}
\usage{
uid(.df = NULL, times = 1, n = NULL, len = 12, prefix = NULL, suffix = NULL)
}
\arguments{
\item{.df}{A data frame object}

\item{times}{Number of times you've called the function so that the seed can be reset}

\item{n}{Length of vector}

\item{len}{length of ID generated}

\item{prefix}{Prefix to add to ID string}

\item{suffix}{Suffix to add to ID string}
}
\value{
A vector or column of unique IDs
}
\description{
Generate unique IDs
}
\examples{
tibble::tibble(a = c(1, 2, 3, 4)) \%>\%
  uid()
}
