% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winning_record.R
\name{winning_record}
\alias{winning_record}
\title{winning_record}
\usage{
winning_record(.df, group, arr = NULL, by = desc)
}
\arguments{
\item{.df}{A data frame object you want to reorder}

\item{group}{Grouping parameter}

\item{arr}{Column to use as the sorting parameter}

\item{by}{Sorting function}
}
\value{
A data frame object with a single row per group
}
\description{
Arranges and chooses the first row of a group based on a sorting parameter
}
\examples{
mpg \%>\% winning_record(group = model, arr = cyl)
}
