% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{write}
\usage{
write(x, path = wd, type = "csv")
}
\arguments{
\item{x}{A data frame object}

\item{path}{Location to write the file to}

\item{type}{Type of file}
}
\value{
A data frame object with the new column in the first position
}
\description{
Function that writes a data frame object to the wd
}
\examples{
.x <- tibble::tibble(a = NA, b = NA)
\dontrun{
write(.x)
}
}
