% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\docType{package}
\name{tabular}
\alias{tabular}
\alias{tabular-package}
\alias{tabular}
\title{tabular: Tabular Class Conversion}
\usage{
tabular(..., class = "data.frame")
}
\arguments{
\item{...}{Arguments passed to the corresponding creation function.}

\item{class}{Desired tabular class (class name or object of that class).}
}
\description{
tabular provides a simple programmatic interface for tabular object conversion. It is meant to simplify the task of writing functions that return tabular objects whose classes match the ones input by the user.

Creates a tabular object of the desired class.
}
\details{
The following tabular classes are supported:

\itemize{
  \item data.frame (see \code{base::\link[base]{data.frame}}):
  \item data.table (see \code{data.table::\link[data.table]{data.table}})
  \item tbl_df (see \code{tibble::\link[tibble]{tibble}})
  \item tbl_dt (see \code{dtplyr::\link[dtplyr]{tbl_dt}})
}
}
\examples{
tabular(a = 1, b = 2)
tabular(a = 1, b = 2, class = "data.frame")
\dontrun{
tabular(a = 1, b = 2, class = "data.table")
tabular(a = 1, b = 2, class = "tbl_df")
tabular(a = 1, b = 2, class = "tbl_dt")
}
}
