% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{tabular_classes}
\alias{tabular_classes}
\alias{col_spec}
\alias{header_node}
\alias{sort_spec}
\alias{pivot_spec}
\alias{derive_spec}
\alias{style_node}
\alias{style_predicate}
\alias{style_spec}
\alias{pagination_spec}
\alias{preset_spec}
\alias{tabular_spec}
\alias{tabular_grid}
\title{tabular S7 classes}
\description{
Internal S7 class definitions for tabular's display-side IR. Users
do not construct these directly except via \code{col_spec()};
everything else is built by the verb chain
(\code{tabular()} -> \code{cols()} -> ... -> \code{emit()}).
}
\keyword{internal}
